Fix extended metadata values type
By default, all initial values for extended metadata are initialized with false
which causes an exception on creating a new series as it complains that
the type does not match the one defined in the catalog.

This patch removes the extended metadata handling methods and reuses the one for the "common" metadata.